move kernels models out of api, fix imports
